pyproject.toml
[project]
name = "ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "firecrawl-py>=2.16.5",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-openai>=0.3.28",
    "langchain-together>=0.3.1",
    "langgraph>=0.6.4",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "together>=1.5.23",
]
